# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Maquettiste editor API
  version: 0.2.0-phase2
  summary: The JSON API the Maquettiste editor's C# functions expose on the editor's own domain (SPEC Section 16), phase 2 subset.
  description: |
    The editor site's `_functions/` answer this API on the editor's domain (`http://maquettiste.localhost:8080` in local
    mode). Every handler is route parsing, the sign-in check and one call into `Maquettiste.Engine`
    (`docs/engineering/engine-design.md` section 15, `docs/engineering/phase2-design.md` section 3). The binding
    design is `docs/engineering/phase2-design.md`; this file is the contract the editor SPA, its generated client and its
    MSW mocks are built from.

    **Serialization.** Every body that mirrors an engine record is that record serialized by `System.Text.Json` with
    `JsonSerializerDefaults.Web` and no custom converters (host-contracts requirement 5): camelCase names, every
    property written, `null` included, enums as the kebab-case strings the engine declares, `TimeSpan` as
    `[-][d.]hh:mm:ss[.fffffff]`, `DateTimeOffset` as ISO 8601 with offset, and non-ASCII and HTML-sensitive characters
    escaped as `\uXXXX` (the default encoder). Computed public properties are serialized too (`ChangeSet.isEmpty`,
    `ValidationReport.hasErrors`). Schemas marked `x-engine-record` name the C# record they mirror; a change to that
    record is a change to this file in the same commit.

...
    and sends `json`, never `element`.

...

...

...

...

...
servers:
...
security:
  - sessionCookie: []
  - editorToken: []
...
paths:
...
  /api/databases/{id}/view:
    parameters:
      - $ref: '#/components/parameters/ElementId'
    get:
      tags: [databases]
      operationId: getDatabaseView
      summary: The resolved tables of one database
      description: |
        The physical model after conventions, mappings and table overlays are applied: every table (synthesized,
        designed or imported), its columns with the attribute each one stores, keys, foreign keys and indexes. The
        Database workspace draws table diagrams from it and the Mappings workspace sets it beside the entity. A model
        with errors returns `view: null` and the diagnostics.
      x-maquettiste-role: viewer
      x-engine-addition: E1 GenerationService.GetDatabaseViewAsync(databaseId)
      responses:
        '200':
          description: The view, or the diagnostics that prevented resolution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseViewResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/databases/{id}/tables:
    parameters:
      - $ref: '#/components/parameters/ElementId'
    get:
      tags: [databases]
      operationId: getDatabaseTables
      summary: The table list of one database, without columns
      description: |
        E5c: a projection of the resolved database for the explorer's Databases and Mappings sections, table-name search and
        the coverage line: every table's key, name, schema, origin, entity or relation and column count, without the columns.
        Unlike `/view` it answers on a model with errors: the tables whose owning elements are valid come back, the rest are
        left out, `partial` is true and the diagnostics say why. The editor keeps the last complete list per database and
        shows it with a stale badge while a response is partial. Per model version the server validates the model once and
        resolves each database asked for on its own (the conceptual layer and that database), so a change costs one database's
        resolve; `diagnostics` then holds the validation diagnostics and that database's resolution diagnostics.
      x-maquettiste-role: viewer
      x-engine-addition: E5c DatabaseTables.GetAsync(databaseId)
      responses:
        '200':
          description: The tables that resolve, the diagnostics, and whether the list is partial.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseTablesResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'
  /api/databases/{id}/tables/{key}:
    parameters:
      - $ref: '#/components/parameters/ElementId'
      - name: key
        in: path
        required: true
        description: The table's key, as `TableSummary.key` (a table file's id, or a synthesized key such as `<entityId>@<databaseId>`).
        schema:
          type: string
    get:
      tags: [databases]
      operationId: getDatabaseTable
      summary: One table of a database with its columns, keys and indexes
      description: |
        E5f: one `TableView` (columns with `attributeId`, primary key, foreign keys, unique constraints, indexes) for a table's
        children in the explorer and column-level "mapped by", without `/view`'s every table. A projection of the same
        per-version resolve as `/tables`: after the first call for a database and model version, a lookup. Like `/tables` it
        answers on a model with errors; `table` is null when no table of the database has the key or when the table is left
        out because an owning element has errors (`partial` then says so).
      x-maquettiste-role: viewer
      x-engine-addition: E5f DatabaseTables.GetTableAsync(databaseId, key)
      responses:
        '200':
          description: The table (or null), the diagnostics, and whether the model has errors.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseTableResult'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthenticated'

  # ------------------------------------------------ reference data and localization (declared; handlers land later)
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mq_session
      description: Set by `POST /api/session`. Host-only, HttpOnly, SameSite=Strict; protected with the host's data-protection keys; carries the user name, role, issue time and a fingerprint of the editor token.
    editorToken:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <MAQUETTISTE_EDITOR_TOKEN>`, the secret site variable. For the CLI, scripts and tests. Compared in fixed time. In local mode a loopback or `MAQUETTISTE_LOCAL_PEERS` caller with a local `Host` and no forwarding header needs neither scheme and is the local admin (see the authentication notes in `info`).'
  parameters:
...
    ElementId:
      name: id
      in: path
      required: true
      description: An element or sub-element id (uppercase ULID).
      schema:
        $ref: '#/components/schemas/Ulid'
      example: 01J92P0V0FJ23CGSNKM7P1W5V7
...
  responses:
...
    Unauthenticated:
      description: No valid credential (`unauthenticated`), or a wrong token (`bad-token`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: Sign in to use the editor.
            status: 401
            code: unauthenticated
    NotFound:
      description: No such element, diagram, plan, path in the plan, job or connection (`not-found`, `not-a-diagram`, `not-a-database`).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: about:blank
            title: No element has the id 01J92P0V0FJ23CGSNKM7P1W5V9.
            status: 404
            code: not-found
...
  schemas:
...
    Ulid:
      description: Uppercase Crockford ULID (engine-design.md D1).
      type: string
      pattern: '^[0-7][0-9A-HJKMNP-TV-Z]{25}$'
...
    Problem:
      description: RFC 9457 problem details (`application/problem+json`), from `Results.Problem`. `code` is stable; `title` and `detail` are for people.
      type: object
      required: [title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          enum:
            - bad-request
            - unsupported-media-type
            - too-large
            - unauthenticated
            - bad-token
            - too-many-attempts
            - forbidden
            - forbidden-origin
            - not-found
            - not-a-diagram
            - not-a-database
            - precondition-required
            - queue-full
            - job-finished
            - model-unavailable
            - internal
        traceId:
          type: string
...
    DiagnosticSeverity:
      type: string
      enum: [error, warning, info]
    Diagnostic:
      x-engine-record: Maquettiste.Engine.Diagnostics.Diagnostic
      type: object
      required: [rule, severity, message, elementId, filePath, jsonPointer, line, column]
      properties:
        rule:
          type: string
          pattern: '^(MQ[1-7][0-9]{3}|x/.+)$'
        severity:
          $ref: '#/components/schemas/DiagnosticSeverity'
        message:
          type: string
        elementId:
          type: [string, 'null']
        filePath:
          type: [string, 'null']
        jsonPointer:
          type: [string, 'null']
        line:
          type: [integer, 'null']
          minimum: 1
        column:
          type: [integer, 'null']
          minimum: 1
...
    DatabaseViewResult:
      x-engine-addition: E1
      x-engine-record: Maquettiste.Engine.DatabaseViewResult
      type: object
      required: [view, diagnostics]
      properties:
        view:
          oneOf:
            - $ref: '#/components/schemas/DatabaseView'
            - type: 'null'
        diagnostics:
          type: array
          items:
            $ref: '#/components/schemas/Diagnostic'
    DatabaseTablesResult:
      x-engine-addition: E5c
      x-engine-record: Maquettiste.Engine.DatabaseTablesResult
      type: object
      required: [tables, diagnostics, partial]
      properties:
        tables:
          type: array
          description: The tables that resolve, in the resolver's order (the same order as `DatabaseView.tables`).
          items:
            $ref: '#/components/schemas/TableSummary'
        diagnostics:
          type: array
          description: Every validation and resolution diagnostic, sorted.
          items:
            $ref: '#/components/schemas/Diagnostic'
        partial:
          type: boolean
          description: |
            The model has errors, so tables whose owning elements have errors (the entity, the relation, the table file or
            overlay, a mapping of the entity or relation, or the database itself) are left out, and the
            rest may change when the errors are fixed. False when the list is complete.
    DatabaseTableResult:
      x-engine-addition: E5f
      x-engine-record: Maquettiste.Engine.DatabaseTableResult
      type: object
      required: [table, diagnostics, partial]
      properties:
        table:
          description: The table, or null when no table of the database has the key or the table is left out (`partial`).
          oneOf:
            - $ref: '#/components/schemas/TableView'
            - type: 'null'
        diagnostics:
          type: array
          description: As `DatabaseTablesResult.diagnostics`.
          items:
            $ref: '#/components/schemas/Diagnostic'
        partial:
          type: boolean
          description: As `DatabaseTablesResult.partial`.
    TableSummary:
      x-engine-addition: E5c
      x-engine-record: Maquettiste.Engine.TableSummary
      type: object
      required: [key, name, schema, origin, entityId, relationId, isJunction, isLookup, columnCount]
      properties:
        key:
          type: string
          description: As `TableView.key`.
        name:
          type: string
        schema:
          type: [string, 'null']
        origin:
          type: string
          enum: [synthesized, designed, imported]
        entityId:
          oneOf:
            - $ref: '#/components/schemas/Ulid'
            - type: 'null'
        relationId:
          oneOf:
            - $ref: '#/components/schemas/Ulid'
            - type: 'null'
        isJunction:
          type: boolean
        isLookup:
          type: boolean
          deprecated: true
          description: Always false. The enum lookup-table storage option is retired (MQ7012); kept so older clients still parse.
        columnCount:
          type: integer
          minimum: 0
    DatabaseView:
      x-engine-addition: E1
      x-engine-record: Maquettiste.Engine.DatabaseView
      type: object
      required: [id, name, dialect, version, defaultSchema, tables]
      properties:
        id:
          $ref: '#/components/schemas/Ulid'
        name:
          type: string
        dialect:
          type: string
        version:
          type: [string, 'null']
        defaultSchema:
          type: [string, 'null']
        tables:
          type: array
          items:
            $ref: '#/components/schemas/TableView'
    TableView:
      x-engine-addition: E1
      x-engine-record: Maquettiste.Engine.TableView
      type: object
      required: [key, name, schema, origin, entityId, relationId, isJunction, isLookup, comment, columns, primaryKey, uniques, foreignKeys, indexes]
      properties:
        key:
          type: string
          description: A table file's id, or a synthesized key such as `<entityId>@<databaseId>` (engine-design.md 7.3); the `elementId` for `sql-ddl`/`table` previews.
        name:
          type: string
        schema:
          type: [string, 'null']
        origin:
          type: string
          enum: [synthesized, designed, imported]
        entityId:
          type: [string, 'null']
        relationId:
          type: [string, 'null']
        isJunction:
          type: boolean
        isLookup:
          type: boolean
          deprecated: true
          description: Always false. The enum lookup-table storage option is retired (MQ7012); kept so older clients still parse.
        comment:
          type: [string, 'null']
        columns:
          type: array
          items:
            $ref: '#/components/schemas/ColumnView'
        primaryKey:
          oneOf:
            - $ref: '#/components/schemas/KeyView'
            - type: 'null'
        uniques:
          type: array
          items:
            $ref: '#/components/schemas/KeyView'
        foreignKeys:
          type: array
          items:
            $ref: '#/components/schemas/ForeignKeyView'
        indexes:
          type: array
          items:
            $ref: '#/components/schemas/IndexView'
    ColumnView:
      x-engine-addition: E1
      x-engine-record: Maquettiste.Engine.ColumnView
      type: object
      required: [key, name, type, nativeType, length, precision, scale, nullable, defaultSql, identity, computed, attributeId, attributePath, isPrimaryKey, isForeignKey, isDiscriminator, position]
      properties:
        key:
          type: string
          description: An attribute path, `discriminator`, `position`, `id`, or a column id.
        name:
          type: string
        type:
          type: string
          description: The built-in type keyword.
        nativeType:
          type: string
          description: The dialect's type, such as `varchar(120)`.
        length:
          type: [integer, 'null']
        precision:
          type: [integer, 'null']
        scale:
          type: [integer, 'null']
        nullable:
          type: boolean
        defaultSql:
          type: [string, 'null']
        identity:
          type: boolean
        computed:
          type: [string, 'null']
        attributeId:
          type: [string, 'null']
        attributePath:
          type: [string, 'null']
        isPrimaryKey:
          type: boolean
        isForeignKey:
          type: boolean
        isDiscriminator:
          type: boolean
        position:
          type: integer
    KeyView:
      x-engine-addition: E1
      x-engine-record: Maquettiste.Engine.KeyView
      type: object
      required: [name, columns]
      properties:
        name:
          type: string
        columns:
          description: Column keys, in key order.
          type: array
          items:
            type: string
    ForeignKeyView:
      x-engine-addition: E1
      x-engine-record: Maquettiste.Engine.ForeignKeyView
      type: object
      required: [name, columns, referencedTable, referencedColumns, onDelete, onUpdate, relationId, endId]
      properties:
        name:
          type: string
        columns:
          type: array
          items:
            type: string
        referencedTable:
          type: string
          description: The referenced table's key.
        referencedColumns:
          type: array
          items:
            type: string
        onDelete:
          type: string
        onUpdate:
          type: string
        relationId:
          type: [string, 'null']
        endId:
          type: [string, 'null']
    IndexView:
      x-engine-addition: E1
      x-engine-record: Maquettiste.Engine.IndexView
      type: object
      required: [name, columns, unique, where]
      properties:
        name:
          type: string
        columns:
          type: array
          items:
            type: object
            required: [column, descending]
            properties:
              column:
                type: string
              descending:
                type: boolean
        unique:
          type: boolean
        where:
          type: [string, 'null']

    # ---------------------------------------------------------------- generation
...